Cleanup 4.1

+++ lab4/zad1/main.c
@@ -9,17 +9,19 @@
 
 pid_t child_pid = 0;
 
-void sigint_handler();
-void sigstp_handler();
-void run_date_printer();
-void kill_date_printer();
-
 void print_current_time(){
   char *time_buff = calloc(50, sizeof(char));
   time_t timestamp = time(NULL);
   strftime(time_buff, 50, "%Y-%m-%d %H:%M:%S", localtime(&timestamp));
   printf("%s\n", time_buff);
   free(time_buff);
+}
+
+void run_date_printer(){
+  if(child_pid != 0) return;
+
+  child_pid = fork();
+  if(child_pid == 0) execl("./spam.sh", "spam", NULL);
 }
 
 void sigstp_handler(){
@@ -34,32 +36,25 @@ void sigstp_handler(){
 
 void sigint_handler(){
   printf("Odebrano sygnal SIGINT.\n");
-  kill(child_pid, SIGKILL);
+  if(child_pid != 0) kill(child_pid, SIGKILL);
   exit(0);
-}
-
-void run_date_printer(){
-  if(child_pid != 0) return;
-
-  child_pid = fork();
-  if(child_pid == 0){
-    execl("./spam.sh", "spam", NULL);
-  }
-  pause();
 }
 
 int main(int argc, char **argv){
 
+  // Handle SIGINT with sigaction
   struct sigaction action;
   action.sa_handler = sigint_handler;
   sigemptyset(&action.sa_mask);
   action.sa_flags = 0;
   sigaction(SIGINT, &action, NULL);
 
+  // Handle SIGSTOP with signal
   signal(SIGTSTP, sigstp_handler);
 
+  // Create data printer process and wait for signals.
   run_date_printer();
-  while(1) {}
+  while(1) pause();
 
   return 0;
 }
